+++ resources/views/Admin/AdminPublic/publics.blade.php
@@ -1,7 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
  <head> 
-  <script type="text/javascript" src="/static/js/jquery-1.8.3.min.js"></script>
   <!-- Required meta tags --> 
   <meta charset="utf-8" /> 
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" /> 
@@ -158,15 +157,14 @@
         </div>
         @endif
       @if(session('success'))
-
-      <a href="javascript:void(0)" class="d-flex align-items-center  btn btn-gradient-danger btn-fw" id="success"> 
+      <div class="d-flex align-items-center  btn btn-gradient-danger btn-fw">
         {{session('success')}}
-      </a>
+      </div>
       @endif
       @if(session('error'))
-      <a href="javascript:void(0)" class="mws-form-message error btn btn-gradient-danger btn-fw" id="error">
+      <div class="mws-form-message error">
           {{session('error')}}
-      </a>
+      </div>
       @endif 
       @section('admin')
      
@@ -185,15 +183,7 @@
     <!-- main-panel ends --> 
    </div> 
    <!-- page-body-wrapper ends --> 
-  </div>
-  <script type="text/javascript">
-    $("#success").click(function(){
-      $("#success").remove();
-    });
-    $("#error").click(function(){
-      $("#error").remove();
-    });
-  </script> 
+  </div> 
   <!-- container-scroller --> 
   <!-- plugins:js --> 
   <script src="/static/admins/vendors/js/vendor.bundle.base.js"></script> 
@@ -206,8 +196,7 @@
   <script src="/static/admins/js/misc.js"></script> 
   <!-- endinject --> 
   <!-- Custom js for this page--> 
-  <script src="/static/admins/js/dashboard.js"></script>
-  <script src="/static/admins/js/core/mws.js"></script>
+  <script src="/static/admins/js/dashboard.js"></script> 
   <!-- End custom js for this page-->   
  </body>
 </html>
